+++ ClassDiagram1.cd
@@ -1,49 +1,54 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <ClassDiagram MajorVersion="1" MinorVersion="1">
   <Class Name="MyGame.Animator">
-    <Position X="3" Y="4" Width="1.5" />
+    <Position X="1.25" Y="4.5" Width="1.5" />
     <TypeIdentifier>
-      <HashCode>AAAAAAAAAABAAAAAAAAAAAABAAIAAAAAwABAIAAAAAA=</HashCode>
-      <FileName>Animator.cs</FileName>
+      <HashCode>AAAAAAAAAABAAAAAAAAAAAABAAAAAAAAwAAAIAAAAAA=</HashCode>
+      <FileName>Animations\Animator.cs</FileName>
     </TypeIdentifier>
   </Class>
   <Class Name="MyGame.Animation">
-    <Position X="4.75" Y="3.75" Width="1.5" />
+    <Position X="3.25" Y="4.5" Width="1.5" />
     <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAAAAAAAACAAAAAIEAAAAQAFAIAAAAAA=</HashCode>
-      <FileName>Animator.cs</FileName>
+      <HashCode>AAAAAAAAAAAAAAAAAAAACAAAAAIEAAAAQAEAIAAAAAA=</HashCode>
+      <FileName>Animations\Animation.cs</FileName>
     </TypeIdentifier>
   </Class>
-  <Class Name="MyGame.Time" Collapsed="true">
-    <Position X="11.25" Y="1.25" Width="1.5" />
+  <Class Name="MyGame.Time">
+    <Position X="9.75" Y="0.75" Width="1.5" />
     <TypeIdentifier>
-      <HashCode>EAAAAAAACCEAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=</HashCode>
-      <FileName>Animator.cs</FileName>
+      <HashCode>EAAAAAAACCEAAAAAAAAAAAAAAAAAAAAAIAAAAAAAAAA=</HashCode>
+      <FileName>Logica\Time.cs</FileName>
     </TypeIdentifier>
   </Class>
   <Class Name="MyGame.Controller">
-    <Position X="8.25" Y="4.25" Width="1.5" />
+    <Position X="9.25" Y="6.5" Width="1.5" />
+    <Compartments>
+      <Compartment Name="Methods" Collapsed="true" />
+    </Compartments>
+    <InheritanceLine Type="MyGame.IController" ManuallyRouted="true" FixedFromPoint="true" FixedToPoint="true">
+      <Path>
+        <Point X="5.688" Y="7.146" />
+        <Point X="5.688" Y="7.655" />
+        <Point X="8.875" Y="7.655" />
+        <Point X="8.875" Y="7.447" />
+        <Point X="9.25" Y="7.447" />
+      </Path>
+    </InheritanceLine>
     <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAAAAAAAACAAAAAIAAAAAQABAAAAAAAA=</HashCode>
-      <FileName>Controller.cs</FileName>
+      <HashCode>AAAAAAAAAAAAAAAAAAAACAAAAAIAAAAAQAAAAAAAAAA=</HashCode>
+      <FileName>MainObject\Controller.cs</FileName>
     </TypeIdentifier>
   </Class>
   <Class Name="MyGame.IController">
-    <Position X="6.5" Y="4" Width="1.5" />
+    <Position X="5" Y="4.5" Width="1.5" />
     <TypeIdentifier>
-      <HashCode>AASAAAAAECAAAAAAAAAQAAQAAAAAAAAAAAAAAAAAAAA=</HashCode>
-      <FileName>IController.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="MyGame.AElement">
-    <Position X="6.25" Y="0.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAQAAAAAACEgAAAgAAAAAAAAAAIAAAAAYABAAAAAAAA=</HashCode>
-      <FileName>Interface.cs</FileName>
+      <HashCode>AACAAAAAEAAAAAAAAAAQAAQAAAIAAAAAQAAAAAAAAAA=</HashCode>
+      <FileName>Logica\IController.cs</FileName>
     </TypeIdentifier>
   </Class>
   <Class Name="MyGame.Hero">
-    <Position X="12.5" Y="3.5" Width="1.5" />
+    <Position X="11.5" Y="4.25" Width="1.5" />
     <NestedTypes>
       <Enum Name="MyGame.Hero.MoveState" Collapsed="true">
         <TypeIdentifier>
@@ -52,23 +57,38 @@
       </Enum>
     </NestedTypes>
     <TypeIdentifier>
-      <HashCode>AAAAAAIAAAAAAAAAAAAAAAAAAAIAAAAAyADAAAAAAAI=</HashCode>
-      <FileName>Player.cs</FileName>
+      <HashCode>AAAAAAIAAAAAAAAAAAAAAAAAAAAAAQAAiACAAAAAAAI=</HashCode>
+      <FileName>MainObject\Hero.cs</FileName>
     </TypeIdentifier>
   </Class>
   <Class Name="MyGame.Player">
-    <Position X="10.5" Y="2.75" Width="1.5" />
+    <Position X="6.75" Y="4.5" Width="1.5" />
+    <InheritanceLine Type="MyGame.ActivElement" ManuallyRouted="true" FixedFromPoint="true" FixedToPoint="true">
+      <Path>
+        <Point X="5.75" Y="3.661" />
+        <Point X="5.75" Y="4.125" />
+        <Point X="7.312" Y="4.125" />
+        <Point X="7.312" Y="4.5" />
+      </Path>
+    </InheritanceLine>
     <TypeIdentifier>
-      <HashCode>AAQAAAAAAAAAAAAAAAIAAAAAAAAAAAAQICAAAAAAAAA=</HashCode>
-      <FileName>Player.cs</FileName>
+      <HashCode>AAAAAAAAAAAAgAAAAAIAAAAAAAAAAAAQICAAAAIAAAA=</HashCode>
+      <FileName>Logica\Player.cs</FileName>
     </TypeIdentifier>
     <Lollipop Position="0.2" />
   </Class>
-  <Interface Name="MyGame.IObject" Collapsed="true">
-    <Position X="8.5" Y="0.75" Width="1.5" />
+  <Class Name="MyGame.ActivElement">
+    <Position X="5" Y="0.5" Width="1.5" />
     <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAAAAAAIAAAAAAAAAAAAQAAAAAAAAAAA=</HashCode>
-      <FileName>Interface.cs</FileName>
+      <HashCode>AAQAAAAAACEgABBhAAAAAAAAAAAAAAAAIAAAAAAAAAA=</HashCode>
+      <FileName>Logica\Interface\ActivElement.cs</FileName>
+    </TypeIdentifier>
+  </Class>
+  <Interface Name="MyGame.IObject">
+    <Position X="7.5" Y="1" Width="1.5" />
+    <TypeIdentifier>
+      <HashCode>AAAAAAAAAAAAgAAAAAIAAAAAQAAAAAAQAAAAAAAAAAA=</HashCode>
+      <FileName>Logica\Interface\IObject.cs</FileName>
     </TypeIdentifier>
   </Interface>
   <Font Name="Segoe UI" Size="9" />
